# Single constructor reflection optimisation