--- a/src/Autofac/Core/Activators/Reflection/ConstructorBinder.cs
+++ b/src/Autofac/Core/Activators/Reflection/ConstructorBinder.cs
@@ -116,9 +116,12 @@ public class ConstructorBinder
     /// <summary>
     /// Get the constructor factory delegate.
     /// </summary>
-    internal Func<object?[], object> GetConstructorInvoker()
+    /// <remarks>Will return null if the constructor contains an invalid parameter.</remarks>
+    internal Func<object?[], object>? GetConstructorInvoker()
     {
-        return FactoryCache.GetOrAdd(Constructor, FactoryBuilder);
+        // Method is only used in zero-parameter constructors, so illegal parameters cannot happen.
+        // Adding the null check + throw just in case.
+        return _factory;
     }
 
     private static Func<object?[], object> GetConstructorInvoker(ConstructorInfo constructorInfo)