--- a/src/Autofac/Core/Activators/Reflection/ReflectionActivator.cs
+++ b/src/Autofac/Core/Activators/Reflection/ReflectionActivator.cs
@@ -98,13 +98,16 @@ public class ReflectionActivator : InstanceActivator, IInstanceActivator
             UseSingleConstructorActivation(pipelineBuilder, binders[0]);
             return;
         }
-        else if (ConstructorSelector is MatchingSignatureConstructorSelector matchingSelector)
+        else if (ConstructorSelector is IConstructorSelectorWithEarlyBinding earlyBindingSelector)
         {
-            var matchedConstructor = matchingSelector.SelectConstructorBinding(binders);
+            var matchedConstructor = earlyBindingSelector.SelectConstructorBinder(binders);
 
-            UseSingleConstructorActivation(pipelineBuilder, matchedConstructor);
+            if (matchedConstructor is not null)
+            {
+                UseSingleConstructorActivation(pipelineBuilder, matchedConstructor);
 
-            return;
+                return;
+            }
         }
 
         _constructorBinders = binders;
@@ -121,10 +124,20 @@ public class ReflectionActivator : InstanceActivator, IInstanceActivator
     {
         if (singleConstructor.ParameterCount == 0)
         {
+            var constructorInvoker = singleConstructor.GetConstructorInvoker();
+
+            if (constructorInvoker is null)
+            {
+                // This is not going to happen, because there is only 1 constructor, that constructor has no parameters,
+                // so there are no conditions under which GetConstructorInvoker will return null in this path.
+                // Throw an error here just in case (and to satisfy nullability checks).
+                throw new NoConstructorsFoundException(_implementationType, string.Format(CultureInfo.CurrentCulture, ReflectionActivatorResources.NoConstructorsAvailable, _implementationType, ConstructorFinder));
+            }
+
             // If there are no arguments to the constructor, bypass all argument binding and pre-bind the constructor.
             var boundConstructor = BoundConstructor.ForBindSuccess(
                 singleConstructor,
-                singleConstructor.GetConstructorInvoker(),
+                constructorInvoker,
                 Array.Empty<Func<object?>>());
 
             // Fast-path to just create an instance.
@@ -212,8 +225,6 @@ public class ReflectionActivator : InstanceActivator, IInstanceActivator
 
     private BoundConstructor[] GetAllBindings(ConstructorBinder[] availableConstructors, IComponentContext context, IEnumerable<Parameter> parameters)
     {
-        // Most often, there will be no `parameters` and/or no `_defaultParameters`; in both of those cases we can avoid allocating.
-        // Do a reference compare with the NoParameters constant; faster than an Any() check for the common case.
         var prioritisedParameters = GetAllParameters(parameters);
 
         var boundConstructors = new BoundConstructor[availableConstructors.Length];
@@ -241,6 +252,8 @@ public class ReflectionActivator : InstanceActivator, IInstanceActivator
 
     private IEnumerable<Parameter> GetAllParameters(IEnumerable<Parameter> parameters)
     {
+        // Most often, there will be no `parameters` and/or no `_defaultParameters`; in both of those cases we can avoid allocating.
+        // Do a reference compare with the NoParameters constant; faster than an Any() check for the common case.
         if (ReferenceEquals(ResolveRequest.NoParameters, parameters))
         {
             return _defaultParameters;