--- a/src/Autofac/Core/Activators/Reflection/ReflectionActivatorResources.Designer.cs
+++ b/src/Autofac/Core/Activators/Reflection/ReflectionActivatorResources.Designer.cs
@@ -19,7 +19,7 @@ namespace Autofac.Core.Activators.Reflection {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class ReflectionActivatorResources {