@jimmycartrette 2022-07-23
> Improvements on the multi-constructor benchmarks (which use the old path) may be down to the use of the `ReferenceEquals` check against the `NoParameters` constant, rather than `Any()`.

Would it be because the multi-contructors still have dependencies, and those dependencies have single constructors which are now optimized?